移动函数位置
