+++ package/default_pkg/function.py
@@ -89,79 +89,3 @@ def YYYY_MM_DD_HH_MM_SS(adjust = "0"):
         return None
     formatted_date_time = now.strftime("%Y_%m_%d - %H-%M-%S")
     return formatted_date_time
-
-def table_task_template_instance(task_instance_list:list, system_pkg:dict):
-    table_list = []
-    heading = ["索引", "版本", "task类型"]
-    table_list.append(heading)
-    index = 0
-    for instance in task_instance_list:
-        table_list.append([str(index), str(instance.version), str(instance.type)])
-        index += 1
-    system_pkg["table_msg"](table_list, heading = True)
-    return None
-
-def create_index_list(input_list:list, adjust = 0):
-    index_list = [i+adjust for i in range(len(input_list))]
-    return index_list
-
-def select_valid_index(user_input:str, select_list:list):
-    """用空格分隔输入，给出所有符合条件的索引，输入空字符串则创建索引列表
-    
-    依赖该文件下函数create_index_list"""
-    # 空字符串输入
-    if user_input == "": return create_index_list(select_list)
-    # 非空字符串输入
-    user_input_list = user_input.split(" ")
-    select_index_list = []
-    for index in user_input_list:
-        try:
-            index = int(index)
-            select_list[index]
-            select_index_list.append(index)
-        except ValueError:
-            continue
-    return select_index_list
-
-def table_compare_task_list(old_task_list:list, new_task_list:list, system_pkg:dict, display = 8):
-    """给出两个默认类型的task_list，进行1.0版本比较
-    
-    display -> int，显示task_list末尾的若干项task"""
-    old_task_list_length = len(old_task_list)
-    new_task_list_length = len(new_task_list)
-    # 获取需显示的项
-    old_list = old_task_list[-display:]
-    new_list = new_task_list[-display:]
-    old_len = len(old_list)
-    new_len = len(new_list)
-    if old_len > new_len: MAX_LEN = old_len
-    else: MAX_LEN = new_len
-    old_list = old_task_list[-MAX_LEN:]
-    new_list = new_task_list[-MAX_LEN:]
-    table_list = []
-    for index in range(-1, -MAX_LEN-1, -1):
-        try:
-            task = old_list[index]
-            display_content = task.content[:8]
-            old_column = [str(old_task_list_length + index), f"{task.date}|{task.attribute}|{display_content}"]
-        except IndexError:
-            old_column = ["", ""]
-        try:
-            task = new_list[index]
-            display_content = task.content[:8]
-            new_column = [str(new_task_list_length + index), f"{task.date}|{task.attribute}|{display_content}"]
-        except IndexError:
-            new_column = ["", ""]
-        table_list.append(old_column + new_column)
-    heading = ["索引", "老版本", "索引", "新版本"]
-    table_list.append(heading)
-    table_list.reverse()
-    system_pkg["table_msg"](table_list, heading = True)
-
-def table_backup_file(backup_list:list, system_pkg:dict):
-    table_list = []
-    heading = ["索引", "路径"]
-    table_list.append(heading)
-    for index, file_path in enumerate(backup_list):
-        table_list.append([str(index), file_path])
-    system_pkg["table_msg"](table_list, heading = True)
